Move entity classes around
